Project_2/chat-service/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
langchain==0.3.27
langchain-core==0.3.72
langchain-google-genai==2.1.8
httpx==0.28.1
pydantic==2.11.7
rapidfuzz==3.13.0
python-dotenv==1.1.1
